switch FFT moments from real to complex
(cherry picked from commit 50dff0b957a72bb4f4df52b14e31ab8ed46cdfc9)


Former-commit-id: c4a2a943e67dc4c71fd49538a3e5916656fe9987

+++ src/62_ctqmc/m_ImpurityOperator.F90
@@ -417,8 +417,8 @@ SUBROUTINE ImpurityOperator_setUmatComplex(this, matU)
                                                                          
   DO iflavor1 = 1, this%flavors                                          
     DO iflavor2 = iflavor1+1, this%flavors                               
-      this%mat_U(iflavor1,iflavor2) = REAL(matU(iflavor1,iflavor2))      
-      this%mat_U(iflavor2,iflavor1) = REAL(matU(iflavor2,iflavor1))      
+      this%mat_U(iflavor1,iflavor2) = matU(iflavor1,iflavor2)      
+      this%mat_U(iflavor2,iflavor1) = matU(iflavor2,iflavor1)      
     END DO                                                               
   END DO                                                                 
 END SUBROUTINE ImpurityOperator_setUmatComplex
